third_party/WebKit/LayoutTests/external/wpt/html/browsers/origin/relaxing-the-same-origin-restriction/document_domain-expected.txt

This is a testharness.js-based test.
PASS sanity checks
PASS current document
FAIL new documentassert_equals: new document has empty domain expected "" but got "web-platform.test"
Harness: the test ran to completion.
